+++ src/primitive.ts
@@ -7,12 +7,12 @@ import { updateState } from "./state";
 /**
  * Always succeed with the value of the argument.
  */
-export const pure = <T>(value: T): Parser<T> =>
+export const pure = <const T>(value: T): Parser<T> =>
     new Parser(state => updateState(state, value));
 
 export const fail = (): Parser<never> =>
     new Parser((state, context) => {
-        context.addError(error.unknown(state.i));
+        context.addError(state.i);
         return null;
     });
 
@@ -21,7 +21,7 @@ export const fail = (): Parser<never> =>
  */
 export const EOI = /* #__PURE__ */ new Parser((state, context) => {
     if (state.i < context.src.length) {
-        context.addError(error.unknown(state.i));
+        context.addError(state.i);
         return null;
     }
     return state;
@@ -37,13 +37,18 @@ export const ANY_EL = /* #__PURE__ */ new Parser((state, context) => {
     if (state.i < context.src.length) {
         return updateState(state, context.src[state.i], 1);
     }
-    context.addError(error.unknown(state.i));
+    context.addError(state.i);
     return null;
 });
 
-export const el = <T>(value: T): Parser<T> => satisfy(srcEl => equals(srcEl, value));
+export const el = <const T>(value: T): Parser<T> =>
+    satisfy(srcEl => equals(srcEl, value), {
+        error: error.expected(
+            typeof value === "string" && value.length === 1 ? value : [value],
+        ),
+    });
 
-export const oneOf = <T>(values: Iterable<T>): Parser<T> => {
+export const oneOf = <const T>(values: Iterable<T>): Parser<T> => {
     const set = new Set<unknown>(values);
     return satisfy(el => set.has(el));
 };
@@ -57,6 +62,7 @@ export const satisfy = <T>(
     f:
         | ((el: unknown, config: Config) => boolean)
         | ((el: unknown, config: Config) => el is T),
+    options?: { error?: error.ParseError },
 ): Parser<T> =>
     new Parser((state, context) => {
         let srcEl: unknown;
@@ -66,21 +72,21 @@ export const satisfy = <T>(
         ) {
             return updateState(state, srcEl, 1);
         }
-        context.addError(error.unknown(state.i));
+        context.addError(state.i, options?.error);
         return null;
     });
 
-export const literal = <T extends Source>(chunk: T): Parser<T> =>
+export const literal = <const T extends Source>(chunk: T): Parser<T> =>
     new Parser((state, context) => {
         if (state.i + chunk.length > context.src.length) {
-            context.addError(error.unknown(state.i));
+            context.addError(state.i, error.expected(chunk));
             return null;
         }
         for (let i = 0; i < chunk.length; i++) {
             const srcEl = context.src[state.i + i];
             const chunkEl = chunk[i];
             if (!equals(srcEl, chunkEl)) {
-                context.addError(error.unknown(state.i + i));
+                context.addError(state.i, error.expected(chunk));
                 return null;
             }
         }