+++ CHANGELOG.md
@@ -3,6 +3,25 @@
 
 ## [Unreleased]
 
+## [0.8.0] - 2023-06-24
+### Added
+- Added `Parser.prototype.apply`
+- Added `Parser.prototype.label`
+- Added `many`
+- Added `manyAccum`
+- Added `string`
+- Added `graphemeString`
+- Added `CODE_POINT`
+- Added `ANY_CHAR`
+- Added `index` and `errors` properties to FailureParseResult.
+
+### Changed
+- Change default for `Parser.prototype.option` to undefined.
+
+### Deprecated
+- `Parser.prototype.many`
+- `Parser.prototype.manyAccum`
+
 ## [0.7.0] - 2023-05-13
 ### Added
 - Added `lookAhead`